Add further funcionality. Rewrite 'insert_recipe' function.
Added additional fields for prep and cooking times, temperature, and an
allergens list. Rewrote the 'insert_recipe' function to allow for
storing the ingredients list and preparation instructions as arrays.
Added a flash message on submiting the form to indicate success.
Added 'showrecipes.html' to display each recipe and its content.