Add delete account functionality
Added the ability for a user to delete their account via a modal
on the Profile page. User is asked to confirm their password
and choose if they wish to also delete all of their own recipes.
If user chooses to delete all, then the recipes are removed from
the database and also from any other user favourites. If not, then
all of the user's recipes as reassigned to the Admin account. Any
recipes in the user's favourites also have their favourite count
decremented by one.